fix(icons): vega icon records
