+++ libs/ui-toolkit/src/components/icon/vega-icons/index.ts
@@ -1,3 +1,3 @@
 export * from './vega-icon';
 export * from './vega-icon.stories';
-export * from './svg';
+export * from './vega-icon-record';
