+++ libs/ui-toolkit/src/components/icon/vega-icons/vega-icon.spec.tsx
@@ -1,10 +1,11 @@
 import { render } from '@testing-library/react';
 
 import { VegaIcon } from './vega-icon';
+import { VegaIconNames } from './vega-icon-record';
 
 describe('VegaIcon', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<VegaIcon name="twitter" />);
+    const { baseElement } = render(<VegaIcon name={VegaIconNames.TWITTER} />);
     expect(baseElement).toBeTruthy();
   });
 });
