+++ libs/ui-toolkit/src/components/icon/vega-icons/vega-icon.stories.tsx
@@ -1,6 +1,7 @@
 import type { Story, Meta } from '@storybook/react';
 import type { VegaIconProps } from './vega-icon';
 import { VegaIcon } from './vega-icon';
+import { VegaIconNames } from './vega-icon-record';
 
 export default {
   component: VegaIcon,
@@ -11,17 +12,17 @@ const Template: Story<VegaIconProps> = (args) => <VegaIcon {...args} />;
 
 export const Default = Template.bind({});
 Default.args = {
-  name: 'edit',
+  name: VegaIconNames.BREAKDOWN,
 };
 
 export const Bigger = Template.bind({});
 Bigger.args = {
-  name: 'deposit',
+  name: VegaIconNames.DEPOSIT,
   size: 8,
 };
 
 export const EvenBigger = Template.bind({});
 EvenBigger.args = {
-  name: 'withdraw',
+  name: VegaIconNames.GLOBE,
   size: 12,
 };
